refactor: remove redundancy in store class by extending the corresponding service
